requirements.txt
python-chess
pyqt5
numpy
